requirements.txt
 python-decouple==3.3
 tweepy==3.8.0
 pymongo==3.10.1